+++ variables.js
@@ -1,0 +1,13 @@
+/* variables */
+export const $arenas = document.querySelector('.arenas'),
+    $fightButton = document.querySelector('.button'),
+    $formFight = document.querySelector('.control'),
+    $chat = document.querySelector('.chat'),
+
+    ATTACK = ['head', 'body', 'foot'],
+
+    HIT = {
+        head: 30,
+        body: 25,
+        foot: 20,
+    };
